Created a helper function for category Post
